autogen.sh: Use glibtoolize if there's no libtoolize
On macOS Apple provides a script called libtool that is unrelated to GNU
libtool (and there is no libtoolize). In order to not conflict with
Apple's libtool, when GNU libtool is installed on macOS it is customary
to install it with a "g" prefix: glibtool and glibtoolize. Update
autogen.sh to find glibtoolize if there is no libtoolize.